wip: moved link preview in to entities
