+++ lib/entities/posts/export.dart
@@ -5,3 +5,4 @@ export 'post_status.dart';
 export 'polls/export.dart';
 export 'medias/export.dart';
 export 'reactions/export.dart';
+export 'link_preview.dart';
