+++ lib/ui/models/export.dart
@@ -1,3 +1,3 @@
-export 'link_preview.dart';
+export '../../entities/posts/link_preview.dart';
 export 'password_security.dart';
 export 'ui_post.dart';
